reorganize mef component imports/exports to prevent circular dependencies
