+++ VSRAD.Package/Commands/BaseBuildWithPreviewCommand.cs
@@ -12,9 +12,9 @@ namespace VSRAD.Package.Commands
     public abstract class BaseBuildWithPreviewCommand : BaseRemoteCommand
     {
         private readonly BuildToolsServer _buildServer;
-        private readonly DTE2 _dte;
         private readonly BuildSteps _buildSteps;
 
+        private DTE2 _dte;
         private BuildEvents _buildEvents;
         private (string localPath, string lineMarker)? _ongoingRun;
 
@@ -27,8 +27,6 @@ namespace VSRAD.Package.Commands
         {
             _buildSteps = buildSteps;
             _buildServer = buildServer;
-            _dte = serviceProvider.GetService(typeof(DTE)) as DTE2;
-            Assumes.Present(_dte);
         }
 
         protected abstract Task<(string localPath, string lineMarker)> ConfigurePreviewAsync();
@@ -36,6 +34,11 @@ namespace VSRAD.Package.Commands
         public override async Task RunAsync()
         {
             await VSPackage.TaskFactory.SwitchToMainThreadAsync();
+            if (_dte == null)
+            {
+                _dte = _serviceProvider.GetService(typeof(DTE)) as DTE2;
+                Assumes.Present(_dte);
+            }
             if (_buildEvents == null)
             {
                 _buildEvents = _dte.Events.BuildEvents;
