+++ VSRAD.Package/ProjectSystem/ToolWindowIntegration.cs
@@ -1,10 +1,10 @@
-﻿using System.ComponentModel.Composition;
+﻿using Microsoft.VisualStudio.ProjectSystem;
+using System.ComponentModel.Composition;
 using VSRAD.Package.Options;
-using VSRAD.Package.ProjectSystem;
 using VSRAD.Package.ProjectSystem.Macros;
 using VSRAD.Package.Server;
 
-namespace VSRAD.Package.ToolWindows
+namespace VSRAD.Package.ProjectSystem
 {
     public delegate void AddWatch(string watch);
 
@@ -21,11 +21,14 @@ namespace VSRAD.Package.ToolWindows
         void AddWatchFromEditor(string watch);
     }
 
+    [Export(typeof(IToolWindowIntegration))]
+    [AppliesTo(Constants.RadOrVisualCProjectCapability)]
     public sealed class ToolWindowIntegration : IToolWindowIntegration
     {
-        public ProjectOptions ProjectOptions { get; }
         public ICommunicationChannel CommunicationChannel { get; }
         public MacroEditManager MacroEditor { get; }
+
+        public ProjectOptions ProjectOptions => _project.Options;
 
         public event AddWatch AddWatch;
 
@@ -35,14 +38,16 @@ namespace VSRAD.Package.ToolWindows
             remove => _debugger.BreakEntered -= value;
         }
 
+        private readonly IProject _project;
         private readonly DebuggerIntegration _debugger;
 
-        public ToolWindowIntegration(ProjectOptions options, ICommunicationChannel channel, MacroEditManager macroEditor, DebuggerIntegration debugger)
+        [ImportingConstructor]
+        public ToolWindowIntegration(IProject project, DebuggerIntegration debugger, ICommunicationChannel channel, MacroEditManager macroEditor)
         {
-            ProjectOptions = options;
+            _project = project;
+            _debugger = debugger;
             CommunicationChannel = channel;
             MacroEditor = macroEditor;
-            _debugger = debugger;
         }
 
         public void AddWatchFromEditor(string watch) => AddWatch(watch);
